requirements.txt
httpx>=0.24.0
tenacity>=8.2.0
playwright>=1.35.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pypdf>=3.10.0
python-pptx>=0.6.21
python-docx>=0.8.11
beautifulsoup4>=4.12.0
streamlit>=1.24.0
pyngrok>=6.0.0
mem0ai>=0.0.12
ollama>=0.1.0
